Render

+++ philo/srcs/get_args/ft_atolu_check.c
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: rbroque <rbroque@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/31 17:19:25 by rbroque           #+#    #+#             */
-/*   Updated: 2023/06/01 18:02:05 by rbroque          ###   ########.fr       */
+/*   Created: 2023/06/14 11:14:02 by rbroque           #+#    #+#             */
+/*   Updated: 2023/09/12 15:47:25 by rbroque          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ static bool	is_inbounds(
 	const unsigned long added_nb
 	)
 {
-	const unsigned long	bound = (ULONG_MAX - added_nb) / 10;
+	const unsigned long	bound = (INT_MAX - added_nb) / 10;
 
 	return (nb <= bound);
 }
